Allow evicting dirty pages via forced eviction to avoid lockups

+++ src/evict/evict_lru.c
@@ -473,7 +473,7 @@ __evict_update_work(WT_SESSION_IMPL *session, bool *eviction_needed)
     /*
      * If application threads are blocked by data in cache, track the fill ratio.
      */
-    uint64_t cache_fill_ratio = bytes_inuse / bytes_max;
+    double cache_fill_ratio = (double)bytes_inuse / (double)bytes_max;
     bool evict_is_hard = LF_ISSET(WT_EVICT_CACHE_HARD);
     if (evict_is_hard) {
         if (cache_fill_ratio < 0.25)
@@ -543,6 +543,7 @@ __evict_update_work(WT_SESSION_IMPL *session, bool *eviction_needed)
     __wt_atomic_store_uint32_relaxed(&evict->flags, flags);
 
     *eviction_needed = F_ISSET(evict, WT_EVICT_CACHE_ANY | WT_EVICT_CACHE_URGENT);
+
     return (0);
 }
